CheckIn:Fix bugzilla bug 154682If the fold "BirtdataCache" is not exist in the system temp fold, cleare the cache will throw NPE

+++ data/org.eclipse.birt.data/src/org/eclipse/birt/data/engine/executor/CacheMapManager.java
@@ -231,6 +231,8 @@ class CacheMapManager
 		private void deleteDir( String dirStr )
 		{
 			File sessionsFolder = new File( dirStr );
+			if ( !sessionsFolder.exists( ) )
+				return;
 			File[] sessionFiles = sessionsFolder.listFiles( );
 			for ( int i = 0; i < sessionFiles.length; i++ )
 				sessionFiles[i].delete( );
